app/requirements.txt
flask==3.0.2
Flask-WTF==1.2.1
Jinja2==3.1.3
google-cloud-storage==2.15.0
google-cloud-firestore==2.15.0
google-cloud-pubsub==2.20.1
google-cloud-trace==1.13.3
google-python-cloud-debugger==4.1
firebase-admin==6.4.0
firebase==4.0.1
# opencensus==0.11.4
